# swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Essa aplicação é parte de um dos requisitos para obtenção do título de bacharel em Ciência da Computação - Centro Universitário Sagrado Coração.",
    "title": "GoLang Products API.",
    "contact": {
      "name": "João Vitor Fraga",
      "email": "jvitor_trabalho@hotmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "paths": {
    "/product": {
      "post": {
        "description": "Método recebe os dados do produtos e realiza o cadastro do mesmo.",
        "tags": [
          "createProduct"
        ],
        "summary": "Realiza a inserção do produto desejado",
        "operationId": "CreateProduct",
        "parameters": [
          {
            "description": "product to add to the list of products",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/productRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "description": "Ao receber um ID, a aplicação retorna o respectivo produto cadastrado no banco.",
        "tags": [
          "getProduct"
        ],
        "summary": "Retorna um produto específico cadastrado",
        "operationId": "GetProduct",
        "parameters": [
          {
            "description": "id of the product",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/productRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      },
      "put": {
        "description": "Ao receber um ID e o JSON contendo os dados a serem alterados, a aplicação atualiza o respectivo produto cadastrado no banco.",
        "tags": [
          "updateProduct"
        ],
        "summary": "Atualiza um produto específico cadastrado",
        "operationId": "UpdateProduct",
        "parameters": [
          {
            "description": "id of the product",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "parameters of a product to be changed",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/productRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      },
      "delete": {
        "description": "Ao receber um ID, a aplicação remove o respectivo produto cadastrado no banco.",
        "tags": [
          "deleteProduct"
        ],
        "summary": "Remove um produto específico",
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "description": "id of the product",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/productRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      }
    },
    "/products": {
      "get": {
        "description": "Todos os produtos persistidos no banco serão retonados.",
        "tags": [
          "getProduct"
        ],
        "summary": "Retornar a lista de produtos cadastrados",
        "operationId": "GetProducts",
        "responses": {
          "200": {
            "$ref": "#/responses/productRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "title": "Product é a estrutura de Produtos responsável por agrupar dados referente ao produto.",
      "properties": {
        "amount": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Amount"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "price": {
          "type": "number",
          "format": "double",
          "x-go-name": "Price"
        },
        "productId": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProductID"
        }
      },
      "x-go-package": "github.com/jotafraga/go-rest-api/models"
    }
  },
  "responses": {
    "badReq": {
      "description": "Error Bad Request",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 400 -  Bad Request",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    },
    "notFoundReq": {
      "description": "Error Not Found",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 404 -  Not Found",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    },
    "okResp": {
      "description": "Success Response",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 200 - OK",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    },
    "productRes": {
      "description": "Product response payload",
      "schema": {
        "$ref": "#/definitions/Product"
      }
    }
  }
}